include: split pcm.h for api & types
Split the pcm.h into 2 files:
pcm.h - exposes the apis
pcm_types.h - types and enums.

This can help to improve the readability.
